Deal with flags values strace doesn't support
These are typically output as a large integer. We obviously can't split this on
'|' and parse like normal. Instead, we set the value of the Flags parsing object
to the string form of this integer argument if the "flags" are a length 1 list
and the one item consists entirely of digits.